# Optimise multiplication