Hi maintainers of rust-num/num-bigint,

I've been hacking on PostgreSQL's numeric.c lately, implementing the Karatsuba algorithm [1].

During this work, I realised that when splitting by half the larger factor's size, then if the smaller factor is less than that, its high part will be zero. I realised that a variable that is zero usually means a formula can be simplified, so I tried to plug in the zero high1 in the formula, and voila, that eliminated a multiplication and most of the additions/subtractions, needing only one multiplication and an addition. This simplification felt too obvious to be novel, have looked around a bit but haven't found it so far. Should be mentioned on the Karatsuba Wikipedia or something I think.

I of course checked also Rust's implementation, and it also seems to benefit from this trick, hence this Pull Request.

Would be fun to know what you think.

Benchmark:

    -test multiply_0           ... bench:          42 ns/iter (+/- 0)
    +test multiply_0           ... bench:          42 ns/iter (+/- 0)

    -test multiply_1           ... bench:       4,271 ns/iter (+/- 68)
    +test multiply_1           ... bench:       4,242 ns/iter (+/- 164)

    -test multiply_2           ... bench:     289,339 ns/iter (+/- 10,532)
    +test multiply_2           ... bench:     288,857 ns/iter (+/- 8,715)

    -test multiply_3           ... bench:     663,387 ns/iter (+/- 17,381)
    +test multiply_3           ... bench:     583,635 ns/iter (+/- 14,997)

    -test multiply_4           ... bench:       7,474 ns/iter (+/- 384)
    +test multiply_4           ... bench:       6,649 ns/iter (+/- 164)

    -test multiply_5           ... bench:      16,376 ns/iter (+/- 254)
    +test multiply_5           ... bench:      13,922 ns/iter (+/- 323)

Thanks for great work!

/Joel

[1] https://www.postgresql.org/message-id/flat/7f95163f-2019-4416-a042-6e2141619e5d@app.fastmail.com