@cuviper 2024-04-17
I suspect the reason this isn't normally considered is that it seems to _add_ a layer of recursive multiplications. That is, this "half" step recurses to two "full" Karatsuba steps with the **same** `x.len()`, though halved `y.len()` in each. Maybe the reduced `y` still works in our favor in further recursion, or perhaps it has less obvious algorithmic effects like cache locality. Your new step would also intercept larger sizes that were headed for the Toom-3 branch, so maybe it's helping more there.

Whatever it is, the benchmarks do favor this change. I'm just trying to think if there are untested scenarios that this will make worse.