@joelonsql 2024-04-18
The pseudo-code in the Wikipedia article on Karatsuba for some reason splits on the longer factor:
```
    /* Calculates the size of the numbers. */
    m = max(size_base10(num1), size_base10(num2))
    m2 = floor(m / 2) 
```

I noticed this code:

```rust
        // When x is smaller than y, it's significantly faster to pick b such that x is split in
        // half, not y:
        let b = x.len() / 2;
```

It seems that, with the Half-Karatsuba step, it doesn't matter if we in the Full-Karatsuba step, split `x` or `y` in half, I get almost identical benchmark results, to a degree that it is hard to tell which is the fastest.

The reason why Full-Karatsuba, without the half-step, when splitting on `y` is slower than `x`, is that one of the multiplications will be meaningless, since the `x0` will be zero.

I think splitting the longer `y` in half-step(s), leads to fewer and more balanced smaller multiplications, than if instead splitting on the shorter `x` and do a full Karatsuba-step.

I've done quite a lot of benchmarks on my PostgreSQL numeric mul_var() patch, but not for `num-bigint`, so it would be good with a benchmark that covers a large range of smaller factors *times* a large range of larger factors, to better understand the effect for different factor lengths and factor length ratios, on different architectures.

In my PostgreSQL patch benchmark, I've measured the performance_ratio as defined as the execution time for the current implementation divided by the execution time for the patched version, by exposing both implementations in the same runtime, to allow increasing the number of executions exponentially, until the performance_ratio converge.

I think it would be fun to work together on developing a very capable benchmark suite for `num-bigint`, if you're up for it. For inspiration, I've attached a plot I created for my PostgreSQL patch, using Dynamic Programming to find the optimal threshold function, which as you can see is not trivial. The threshold line goes between the blue and magenta areas. The black line segment is a manually created threshold line, which I think captures the interesting performance gain area, without causing any significant performance regressions on the most prioritised architectures.

![image](https://github.com/rust-num/num-bigint/assets/1784776/b793c1f6-c6db-4a3e-830c-613fee23f42f)

**Important:** The above plot is **not** for this patch, it's from my separate PostgreSQL patch, included here just for inspiration.