@joelonsql 2024-04-22
Hi @cuviper,

I've now done benchmarks across varying sizes of BigUint. Here's what I found: when the factor sizes are identical, there's no noticeable difference in the performance ratio. It's when we introduce a length disparity between factors that the performance ratio starts to climb, highlighting the Half Karatsuba's efficiency in handling size asymmetry.

Interestingly, keeping the larger factor constant and varying the smaller factor shows that the performance_ratio increases with the length disparity, but only to a certain point. Beyond this point, it begins to taper off, until it approaches 1.0 when factor sizes are equal.

This seems encouraging. Eager to hear your thoughts.

The benchmark below was produced using [https://github.com/joelonsql/rust-timeit/blob/master/examples/num_bigint_half_karatsuba.rs](url)
![image](https://github.com/rust-num/num-bigint/assets/1784776/feca3730-06a2-4940-bcaa-b9e92f808871)

As shown in the plot, `b_exp` and `c_exp` have been varied from from 1 to 17.