@joelonsql 2024-04-22
Another benchmark with factors up to `1 << 24` bit_size

![image](https://github.com/rust-num/num-bigint/assets/1784776/c4c204b5-4417-4595-be02-3e9ea5aa0ebe)