@joelonsql 2024-04-22
Here is another benchmark with 256 x 256 measurements, on a linear scale, and with the number of big_digits as the axes.

![image](https://github.com/rust-num/num-bigint/assets/1784776/74f97505-2aa7-47eb-bc7b-f6169aca16f4)