@cuviper 2024-04-24
That's **incredibly** thorough, thank you! I am quite satisified. :smile: 

Another thing about the anomalies -- one of the first things that `mac3` does is swap args to ensure `x.len() <= y.len()`. So we really should be mirrored along that `y = x` diagonal, and the few points that look otherwise are almost certainly noise. For example, it looks like that purple point is your (432, 688), but if it were real it should have been similar at (688, 432). I'm glad your reruns also support this, or else it would seem rather strange...